+++ lib/clixon/clixon_validate.h
@@ -43,7 +43,7 @@
 /*
  * Prototypes
  */
-int xml_yang_validate_rpc(clicon_handle h, cxobj *xrpc, cxobj **xret);
+int xml_yang_validate_rpc(clicon_handle h, cxobj *xrpc, int expanddefault, cxobj **xret);
 int xml_yang_validate_rpc_reply(clicon_handle h, cxobj *xrpc, cxobj **xret);
 int xml_yang_validate_add(clicon_handle h, cxobj *xt, cxobj **xret);
 int xml_yang_validate_list_key_only(cxobj *xt, cxobj **xret);